v8.3

+++ src/base/task/thread_pool/initialization_util.h
@@ -1,4 +1,4 @@
-// Copyright 2016 The Chromium Authors. All rights reserved.
+// Copyright 2016 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -17,7 +17,7 @@ BASE_EXPORT size_t
 RecommendedMaxNumberOfThreadsInThreadGroup(size_t min,
                                            size_t max,
                                            double cores_multiplier,
-                                           size_t offset);
+                                           int offset);
 
 }  // namespace base
