+++ src/chrome/browser/web_applications/preinstalled_web_apps/preinstalled_web_app_definition_utils.cc
@@ -1,9 +1,10 @@
-// Copyright 2020 The Chromium Authors. All rights reserved.
+// Copyright 2020 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 #include "chrome/browser/web_applications/preinstalled_web_apps/preinstalled_web_app_definition_utils.h"
 
+#include "base/numerics/safe_conversions.h"
 #include "chrome/browser/browser_process.h"
 #include "third_party/icu/source/common/unicode/localematcher.h"
 #include "ui/base/resource/resource_bundle.h"
@@ -55,7 +56,7 @@ const char* GetTranslatedName(const char* utf8_default_name,
   if (best_index == -1)
     return utf8_default_name;
 
-  return translations[best_index].utf8_translation;
+  return translations[base::checked_cast<size_t>(best_index)].utf8_translation;
 }
 
 }  // namespace web_app
