+++ src/components/services/heap_profiling/allocation.h
@@ -1,4 +1,4 @@
-// Copyright 2019 The Chromium Authors. All rights reserved.
+// Copyright 2019 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -45,7 +45,7 @@ struct AllocationSite {
   };
 
  private:
-  const uint32_t hash_;
+  const size_t hash_;
 };
 
 inline bool operator==(const AllocationSite& a, const AllocationSite& b) {
