+++ src/media/base/color_plane_layout.cc
@@ -1,14 +1,16 @@
-// Copyright 2019 The Chromium Authors. All rights reserved.
+// Copyright 2019 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 #include "media/base/color_plane_layout.h"
 
+#include "base/check_op.h"
+
 namespace media {
 
 ColorPlaneLayout::ColorPlaneLayout() = default;
 
-ColorPlaneLayout::ColorPlaneLayout(int32_t stride, size_t offset, size_t size)
+ColorPlaneLayout::ColorPlaneLayout(size_t stride, size_t offset, size_t size)
     : stride(stride), offset(offset), size(size) {}
 
 ColorPlaneLayout::~ColorPlaneLayout() = default;
