+++ src/media/mojo/common/mojo_decoder_buffer_converter.h
@@ -1,4 +1,4 @@
-// Copyright 2016 The Chromium Authors. All rights reserved.
+// Copyright 2016 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -8,7 +8,7 @@
 #include <memory>
 
 #include "base/containers/circular_deque.h"
-#include "base/memory/ref_counted.h"
+#include "base/memory/scoped_refptr.h"
 #include "media/base/demuxer_stream.h"
 #include "media/mojo/mojom/media_types.mojom.h"
 #include "mojo/public/cpp/system/data_pipe.h"
@@ -98,7 +98,7 @@ class MojoDecoderBufferReader {
   base::OnceClosure flush_cb_;
 
   // Number of bytes already read into the current buffer.
-  uint32_t bytes_read_;
+  size_t bytes_read_;
 };
 
 // Converts media::DecoderBuffers to mojom::DecoderBuffers, writing the data
@@ -152,7 +152,7 @@ class MojoDecoderBufferWriter {
   base::circular_deque<scoped_refptr<DecoderBuffer>> pending_buffers_;
 
   // Number of bytes already written from the current buffer.
-  uint32_t bytes_written_;
+  size_t bytes_written_;
 };
 
 }  // namespace media
