+++ src/mojo/core/handle_table.h
@@ -1,4 +1,4 @@
-// Copyright 2014 The Chromium Authors. All rights reserved.
+// Copyright 2014 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -122,7 +122,7 @@ class MOJO_SYSTEM_IMPL_EXPORT HandleTable
 
   base::Lock lock_;
 
-  uint64_t next_available_handle_ = 1;
+  uintptr_t next_available_handle_ = 1;
 };
 
 }  // namespace core
